Move HotelSearchUIIntegrationTests assertions to a separated extension file
